refactor: change the study components folder
