+++ src/pages/home/deck-section.vue
