+++ src/pages/home/study-cards.vue
