+++ src/pages/index.ts
@@ -1,2 +1,1 @@
 export { default as HomePage } from "./home/home.vue";
-export { default as StudyAreaPage } from "./study-area/study-area.vue";
